Implemented email system with EthernaRCL Removed EthernaSSO.RCL
