+++ EthernaSSO.sln
@@ -37,10 +37,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "workflows", "workflows", "{
 		.github\workflows\unstable-release.yml = .github\workflows\unstable-release.yml
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "resources", "resources", "{1F3590A3-E673-48CB-BC3D-7EF9671EE42B}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EthernaSSO.RCL", "src\EthernaSSO.RCL\EthernaSSO.RCL.csproj", "{E609F100-1AF8-4F48-8F17-7487DD8BE392}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,10 +59,6 @@ Global
 		{8EE2E2BE-2FB6-4E96-8F34-08F9CDACFCFE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EE2E2BE-2FB6-4E96-8F34-08F9CDACFCFE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8EE2E2BE-2FB6-4E96-8F34-08F9CDACFCFE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E609F100-1AF8-4F48-8F17-7487DD8BE392}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E609F100-1AF8-4F48-8F17-7487DD8BE392}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E609F100-1AF8-4F48-8F17-7487DD8BE392}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E609F100-1AF8-4F48-8F17-7487DD8BE392}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -79,8 +71,6 @@ Global
 		{AF83D80A-77BE-48C8-ABE3-D1373F20F2B0} = {5AE428C2-CD56-4E51-A229-3F4C137D9D83}
 		{8EE2E2BE-2FB6-4E96-8F34-08F9CDACFCFE} = {BB9B2543-8741-4D95-AC15-35923C422E60}
 		{61C8B690-EEA5-440E-B5B5-7C8D5098503C} = {1C01CDAC-366E-4F9B-8A52-25F7770FE92A}
-		{1F3590A3-E673-48CB-BC3D-7EF9671EE42B} = {A0B68B53-E0DF-4EC9-AFAA-7476C8A26256}
-		{E609F100-1AF8-4F48-8F17-7487DD8BE392} = {1F3590A3-E673-48CB-BC3D-7EF9671EE42B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6EB5F2B3-8912-4D50-BAB1-C3884E3BF3C0}
