+++ src/EthernaSSO.Domain/EthernaSSO.Domain.csproj
@@ -14,6 +14,7 @@
 
   <ItemGroup>
     <PackageReference Include="Etherna.DomainEvents" Version="1.3.0" />
+    <PackageReference Include="EthernaRCL" Version="0.1.0-alpha.15" />
     <PackageReference Include="Microsoft.AspNetCore.Authentication.Etherna" Version="0.3.0-alpha.7" />
     <PackageReference Include="Microsoft.Extensions.Identity.Core" Version="5.0.12" />
     <PackageReference Include="MongODM.Core" Version="0.23.0-alpha.37" />
