--- src/EthernaSSO.Domain/Helpers/EmailHelper.cs
@@ -1,46 +1,0 @@
-﻿//   Copyright 2021-present Etherna Sagl
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-
-using System;
-using System.Globalization;
-using System.Text.RegularExpressions;
-
-namespace Etherna.SSOServer.Domain.Helpers
-{
-    public static class EmailHelper
-    {
-        // Consts.
-        public const string EmailRegex = @"\A(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?)\Z";
-
-        // Static methods.
-        public static bool IsValidEmail(string email) =>
-            Regex.IsMatch(email, EmailRegex, RegexOptions.IgnoreCase);
-
-        public static string NormalizeEmail(string email)
-        {
-            if (email is null)
-                throw new ArgumentNullException(nameof(email));
-
-            email = email.ToUpper(CultureInfo.InvariantCulture); //to upper case
-
-            var components = email.Split('@');
-            var username = components[0];
-            var domain = components[1];
-
-            var cleanedUsername = username.Split('+')[0]; //remove chars after '+' symbol, if present
-
-            return $"{cleanedUsername}@{domain}";
-        }
-    }
-}
