+++ src/EthernaSSO.Services/EthernaSSO.Services.csproj
@@ -1,7 +1,8 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
 
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
+    <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
     <RootNamespace>Etherna.SSOServer.Services</RootNamespace>
 
     <Authors>Etherna Sagl</Authors>
@@ -20,12 +21,10 @@
     <PackageReference Include="Etherna.DomainEvents.AspNetCore" Version="1.3.0" />
     <PackageReference Include="IdentityServer4.AspNetIdentity" Version="4.1.2" />
     <PackageReference Include="MongODM.Hangfire" Version="0.23.0-alpha.37" />
-    <PackageReference Include="SendGrid" Version="9.24.4" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\EthernaSSO.Domain\EthernaSSO.Domain.csproj" />
-    <ProjectReference Include="..\EthernaSSO.RCL\EthernaSSO.RCL.csproj" />
   </ItemGroup>
 
 </Project>
