+++ src/EthernaSSO.Services/Views/Emails/ConfirmEmailChange.cshtml
