+++ src/EthernaSSO.Services/Views/Emails/_ViewImports.cshtml
@@ -13,4 +13,4 @@
     *   limitations under the License.
 *@
 
-@namespace Etherna.SSOServer.RCL.Views.Emails
+@namespace Etherna.SSOServer.Services.Views.Emails
