+++ src/EthernaSSO.Services/Views/Emails/_ViewStart.cshtml
@@ -14,5 +14,5 @@
 *@
 
 @{
-    Layout = "EmailLayout";
+    Layout = "../../Views/Shared/EmailLayout.cshtml"; //ethernaRcl
 }
