+++ src/EthernaSSO/Areas/Admin/Pages/Invitations/Index.cshtml.cs
@@ -12,12 +12,12 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
+using Etherna.RCL.Helpers;
+using Etherna.RCL.Services;
 using Etherna.SSOServer.Configs;
 using Etherna.SSOServer.Domain;
-using Etherna.SSOServer.Domain.Helpers;
 using Etherna.SSOServer.Domain.Models;
-using Etherna.SSOServer.RCL.Views.Emails;
-using Etherna.SSOServer.Services.Utilities;
+using Etherna.SSOServer.Services.Views.Emails;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
