+++ src/EthernaSSO/Areas/Api/Services/IdentityControllerService.cs
@@ -12,6 +12,7 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
+using Etherna.RCL.Helpers;
 using Etherna.SSOServer.Areas.Api.DtoModels;
 using Etherna.SSOServer.Domain;
 using Etherna.SSOServer.Domain.Helpers;
