+++ src/EthernaSSO/Startup.cs
@@ -17,6 +17,7 @@ using Etherna.MongODM;
 using Etherna.MongODM.AspNetCore.UI;
 using Etherna.MongODM.Core.Options;
 using Etherna.RCL.Exceptions;
+using Etherna.RCL.Settings;
 using Etherna.SSOServer.Configs;
 using Etherna.SSOServer.Configs.Hangfire;
 using Etherna.SSOServer.Configs.Identity;
@@ -27,6 +28,7 @@ using Etherna.SSOServer.Domain.Models;
 using Etherna.SSOServer.Extensions;
 using Etherna.SSOServer.Persistence;
 using Etherna.SSOServer.Persistence.Settings;
+using Etherna.SSOServer.Services;
 using Etherna.SSOServer.Services.Settings;
 using Etherna.SSOServer.Services.Tasks;
 using Hangfire;
